requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
numpy==1.26.4
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdfplumber==0.11.0
python-dotenv==1.0.0
pydantic==2.5.3
sentence-transformers==2.3.1
pinecone==3.2.2
groq==0.4.1
